refactor tenderly command
