+++ cypress/support/constants.js
@@ -1,0 +1,35 @@
+const BIG_AMOUNT = '11579208923731619542357098500868790785326998466564056'
+const NATIVE_TOKEN_ADDRESS = '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee'
+const FROM_WALLET_ADDRESS = '0x0193a8a52d77e27bdd4f12e0cdd52d8ff1d97d68'
+const MAX_UINT = '10000000000000000000000000000000000000000000000000000000000000000000'
+const ZAP_ROUTER_ADDESS = '0x30c5322e4e08ad500c348007f92f120ab4e2b79e'
+
+const Network = {
+  MAINNET: 1,
+  BSC: 56,
+  POLYGON: 137,
+  ZKEVM: 1101,
+  AVALANCHE: 43114,
+  FANTOM: 250,
+  ARBITRUM: 42161,
+  OPTIMISM: 10,
+  BASE: 8453,
+}
+const Holders = {
+  [Network.MAINNET]: {},
+  [Network.POLYGON]: {},
+  [Network.ARBITRUM]: {
+    '0xaf88d065e77c8cC2239327C5EDb3A432268e5831': '0xa843392198862f98d17e3aa1421b08f2c2020cff',
+  },
+}
+
+module.exports = {
+  NATIVE_TOKEN_ADDRESS,
+  FROM_WALLET_ADDRESS,
+  BIG_AMOUNT,
+  MAX_UINT,
+  ZAP_ROUTER_ADDESS,
+  Network,
+  Holders,
+}
+console.log(Holders[42161]['0xaf88d065e77c8cC2239327C5EDb3A432268e5831'])
